requirements.txt
numpy==1.23.5
tensorflow==2.12.0
matplotlib==3.7.1
keras==2.12.0
scikit-learn==1.2.2
Pillow==9.5.0
scipy==1.10.1